My Expense Report v3
This commit captures 6 weeks of hacking as I learn the android
platform. The User Interface is quite different from the code I forked.

The database format has changed and installing this app performs a
non-reversible upgrade from the original data format. You will have
to delete all the data to re-install the original app.

Built with IDE Android Studio 2 and Gradle 2.

This app can not be uploaded to Google Play because the app id is not
changed. This version must be side-loaded over usb.

WARNING: if you have the original simeonge app installed, you will not
be able to downgrade back to it.

Tested on API 22.

Signed-off-by: Mike Scheutzow <mjs973@gmail.com>